\name{umf-package}
\alias{umf-package}
\alias{umf}
\docType{package}
\title{
\packageTitle{umf}
}
\description{
\packageDescription{umf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{umf}
\packageIndices{umf}
}
\author{
\packageAuthor{umf}

Maintainer: \packageMaintainer{umf}
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
