\name{umf_Classes-Methods}
\alias{umf_Classes-Methods}
\title{
umf_Classes-Methods
}
\description{
Creates 'Ussing' class and defines methods for plot(), print(), and summary() functions. Defines new generic method view.plots() which allows for viewing on indiviual plots one-by-one.
}

\author{
Eric S. Davis
}
